# book-network-api/src/main/java/com/alibou/book/book/BookResponse.java
package com.alibou.book.book;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BookResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private String synopsis;
    private String owner;

    private byte[] cover;
    private double rate;
    private boolean archived;
    private boolean shareable;
}