# book-network-api/src/main/java/com/alibou/book/book/BorrowedBookResponse.java
package com.alibou.book.book;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;

    private double rate;
    private boolean returned;
    private boolean returnApproved;
}