# book-network-api/src/main/java/com/alibou/book/feedback/FeedBackRequest.java
package com.alibou.book.feedback;

import jakarta.validation.constraints.*;

public record FeedBackRequest(
        @Positive(message = "200")
        @Min(value = 0, message = "201")
        @Max(value = 5, message = "202")
        Double note,

        @NotNull(message = "203")
        @NotEmpty(message = "203")
        @NotBlank(message = "203")
        String comment,

        @NotNull(message = "2°4")
        Integer bookId) {
}